feat: implement observer camera position overlay and integrate with player camera manager
