+++ src/app/managers/player-camera-position-manager.ts
@@ -3,11 +3,11 @@ import { PlayerManager } from '../player/player-manager';
 import { debugPrint } from '../utils/debug-print';
 import { EventEmitter } from '../utils/events/event-emitter';
 import { EVENT_ON_PLAYER_LEFT, EVENT_ON_PRE_MATCH } from '../utils/events/event-constants';
-import { CreateObserverButton } from '../utils/observer-helper';
 import { NameManager } from './names/name-manager';
 import { GlobalGameData } from '../game/state/global-game-state';
 import { MinimapIconManager } from './minimap-icon-manager';
 import { SettingsContext } from '../settings/settings-context';
+import { ObserverCameraPositionOverlay } from '../triggers/visuals/observer-camera-position-overlay';
 
 export type CamPositionData = {
 	x: number;
@@ -22,9 +22,7 @@ export default class PlayerCameraPositionManager {
 	private displayPositionData: Map<player, CamPositionData> = new Map<player, CamPositionData>();
 	private frames: Map<player, { box: framehandle; text: framehandle; minimapIcon: framehandle }> = new Map();
 	private syncTrigger: trigger;
-	private overlayVisible: boolean = false;
-	private toggleButton: framehandle;
-	private toggleIcon: framehandle;
+	private observerCameraPositionOverlay: ObserverCameraPositionOverlay;
 
 	public static getInstance() {
 		if (this.instance === undefined) {
@@ -45,15 +43,20 @@ export default class PlayerCameraPositionManager {
 
 		TriggerAddAction(this.syncTrigger, () => this.onSync());
 
+		this.observerCameraPositionOverlay = ObserverCameraPositionOverlay.getInstance();
+		this.observerCameraPositionOverlay.onVisibilityChanged((isVisible) => {
+			if (!isVisible) {
+				this.hidePlayerFrames();
+			}
+		});
+
 		EventEmitter.getInstance().on(EVENT_ON_PLAYER_LEFT, (player) => {
 			this.removePlayerFrame(player.getPlayer());
 		});
 
 		EventEmitter.getInstance().on(EVENT_ON_PRE_MATCH, () => {
-			this.updateEligibility();
+			this.observerCameraPositionOverlay.updateEligibility();
 		});
-
-		this.createToggleButton();
 
 		// Network sync timer — syncs position every 0.5s
 		const syncTimer = CreateTimer();
@@ -90,7 +93,7 @@ export default class PlayerCameraPositionManager {
 
 	private isOverlayVisibleForPlayer(p: player): boolean {
 		if (IsPlayerObserver(p) || EDITOR_DEVELOPER_MODE) {
-			return this.overlayVisible;
+			return this.observerCameraPositionOverlay.isOverlayVisible();
 		}
 
 		const localActivePlayer = PlayerManager.getInstance().players.get(p);
@@ -99,74 +102,6 @@ export default class PlayerCameraPositionManager {
 		}
 
 		return false;
-	}
-
-	private updateEligibility() {
-		const localPlayer = GetLocalPlayer();
-
-		BlzFrameSetVisible(this.toggleButton, false);
-		BlzFrameSetVisible(this.toggleIcon, false);
-		if (IsPlayerObserver(localPlayer) || EDITOR_DEVELOPER_MODE) {
-			BlzFrameSetVisible(this.toggleButton, true);
-			BlzFrameSetVisible(this.toggleIcon, true);
-		} else {
-			BlzFrameSetEnable(this.toggleButton, false);
-		}
-	}
-
-	private createToggleButton(): void {
-		const gameUI = BlzGetOriginFrame(ORIGIN_FRAME_GAME_UI, 0);
-		const ctx = 500; // unique context to avoid collision with player frames
-
-		this.toggleIcon = BlzCreateFrameByType('BACKDROP', 'CamToggleIcon', gameUI, '', ctx);
-		BlzFrameSetPoint(this.toggleIcon, FRAMEPOINT_TOPLEFT, gameUI, FRAMEPOINT_TOPLEFT, 0.138, -0.025);
-		BlzFrameSetSize(this.toggleIcon, 0.02, 0.02);
-		BlzFrameSetTexture(this.toggleIcon, 'ReplaceableTextures\\CommandButtonsDisabled\\DISBTNTelescope.blp', 0, true);
-
-		// Use ScriptDialogButton — same template as the leaderboard button, proven to work with observer hover detection
-		this.toggleButton = BlzCreateFrameByType('GLUETEXTBUTTON', 'CamToggleButton', gameUI, 'ScriptDialogButton', ctx);
-		BlzFrameSetPoint(this.toggleButton, FRAMEPOINT_TOPLEFT, gameUI, FRAMEPOINT_TOPLEFT, 0.138, -0.025);
-		BlzFrameSetSize(this.toggleButton, 0.02, 0.02);
-		BlzFrameSetText(this.toggleButton, '');
-		BlzFrameSetAlpha(this.toggleButton, 0);
-
-		const localPlayer = GetLocalPlayer();
-
-		// Only visible for observers or developers
-		BlzFrameSetVisible(this.toggleButton, false);
-		BlzFrameSetVisible(this.toggleIcon, false);
-		if (IsPlayerObserver(localPlayer) || EDITOR_DEVELOPER_MODE) {
-			BlzFrameSetVisible(this.toggleButton, true);
-			BlzFrameSetVisible(this.toggleIcon, true);
-		} else {
-			BlzFrameSetEnable(this.toggleButton, false);
-		}
-
-		// Observer hover-click — same pattern as leaderboard button in ranked-statistics-view
-		CreateObserverButton(this.toggleButton, IsPlayerObserver(localPlayer), () => {
-			this.toggleOverlay();
-		});
-	}
-
-	private toggleOverlay(): void {
-		this.overlayVisible = !this.overlayVisible;
-		BlzFrameSetText(this.toggleButton, '');
-
-		const texture = this.overlayVisible
-			? 'ReplaceableTextures\\CommandButtons\\BTNTelescope.blp'
-			: 'ReplaceableTextures\\CommandButtonsDisabled\\DISBTNTelescope.blp';
-		BlzFrameSetTexture(this.toggleIcon, texture, 0, true);
-
-		if (!this.overlayVisible) {
-			this.frames.forEach((frame) => {
-				BlzFrameSetVisible(frame.box, false);
-				BlzFrameSetVisible(frame.text, false);
-				BlzFrameSetVisible(frame.minimapIcon, false);
-			});
-		}
-
-		BlzFrameSetEnable(this.toggleButton, false);
-		BlzFrameSetEnable(this.toggleButton, true);
 	}
 
 	private createPlayerFrame(p: player): { box: framehandle; text: framehandle; minimapIcon: framehandle } {
@@ -298,11 +233,7 @@ export default class PlayerCameraPositionManager {
 		const localPlayer = GetLocalPlayer();
 
 		if (!this.isEligiblePlayer(localPlayer) || !this.isOverlayVisibleForPlayer(localPlayer) || GlobalGameData.matchState !== 'inProgress') {
-			this.frames.forEach((frame) => {
-				BlzFrameSetVisible(frame.box, false);
-				BlzFrameSetVisible(frame.text, false);
-				BlzFrameSetVisible(frame.minimapIcon, false);
-			});
+			this.hidePlayerFrames();
 			return;
 		}
 
@@ -337,6 +268,14 @@ export default class PlayerCameraPositionManager {
 		const visLen = this.visibleLength(name);
 		BlzFrameSetSize(frame.text, Math.max(0.02, visLen * 0.005 + 0.01), 0.0058);
 		BlzFrameSetText(frame.text, name);
+	}
+
+	private hidePlayerFrames(): void {
+		this.frames.forEach((frame) => {
+			BlzFrameSetVisible(frame.box, false);
+			BlzFrameSetVisible(frame.text, false);
+			BlzFrameSetVisible(frame.minimapIcon, false);
+		});
 	}
 
 	// Returns the number of visible characters, stripping |cFFRRGGBB (10 chars) and |r (2 chars)
